+++ R/datasimzinb.R
@@ -14,7 +14,7 @@ zinb.simdata <- function(n,p,B, mu,mu.nois,theta,pi){
 # create "adjacency" matrix A from the adjacency matrix B
     if(nrow(B) != ncol(B)){
       print("not a symmetric matrix")
-      return
+      return()
     }
     A <- diag(1,nrow=nrow(B),ncol=ncol(B))
     for(i in 1:(nrow(B)-1)){
@@ -33,10 +33,10 @@ zinb.simdata <- function(n,p,B, mu,mu.nois,theta,pi){
    nonzero.sigma <- ltri.sigma[which(ltri.sigma !=0 )]
    Y.mu <- c(rep(mu,nrow(sigma)), nonzero.sigma)
 ##data matrix
-   Y <-  matrix(unlist(do.call(rbind, parallel::mclapply(Y.mu,function(i) {rzinbinom(n,mu=i,theta, pi=pi)}))),length(Y.mu),n)
+   Y <-  matrix(unlist(do.call(rbind, lapply(Y.mu,function(i) {rzinbinom(n,mu=i,theta, pi=pi)}))),length(Y.mu),n)
    X <- A%*%Y
    # add noise to all the nodes.
-   X <- X + matrix(unlist(do.call(rbind, parallel::mclapply(rep(mu.nois,p),function(i) { rzinbinom(n,mu=i,theta=1, pi=pi)}))),p,n)
+   X <- X + matrix(unlist(do.call(rbind, lapply(rep(mu.nois,p),function(i) { rzinbinom(n,mu=i,theta=1, pi=pi)}))),p,n)
 
    return(t(X))
 }